tests/wpt/web-platform-tests/speculation-rules/prerender/response-code-non-successful.html

<!DOCTYPE html>
<title>Check that non-successful responses result in discarding the prerender</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="resources/utils.js"></script>
<body>
<script>

setup(() => assertSpeculationRulesIsSupported());

function test_prerender_response_code(code) {
  promise_test(async t => {
    const {exec, tryToActivate} = await create_prerendered_page(t, {code});
    const result = await tryToActivate();
    assert_equals(result, 'discarded');
  },`Responses with code ${code} should be discarded`);
}

for (const code of [204, 205, 402, 404, 500, 503]) {
  test_prerender_response_code(code);
}

</script>
</body>
